@@ -22,11 +22,11 @@
     }
   },
   "require": {
-    "php": "^7.1 || ^8.0",
-    "ext-json": "*",
-    "berlioz/helpers": "^1.0"
+    "php": "^8.0",
+    "berlioz/helpers": "^1.0",
+    "colinodell/json5": "^2.2"
   },
   "require-dev": {
-    "phpunit/phpunit": "^7.5 || ^8.0 || ^9.0"
+    "phpunit/phpunit": "^9.3"
   }
 }
