@@ -1,9 +1,9 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2017 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
@@ -22,18 +22,18 @@ interface ConfigAwareInterface
     /**
      * Get config.
      *
-     * @return \Berlioz\Config\ConfigInterface|null
+     * @return ConfigInterface|null
      */
     public function getConfig(): ?ConfigInterface;
 
     /**
      * Set config.
      *
-     * @param \Berlioz\Config\ConfigInterface $config
+     * @param ConfigInterface $config
      *
      * @return static
      */
-    public function setConfig(ConfigInterface $config);
+    public function setConfig(ConfigInterface $config): static;
 
     /**
      * Has config?
