@@ -1,16 +1,14 @@
 <?php
-/**
+/*
  * This file is part of Berlioz framework.
  *
  * @license   https://opensource.org/licenses/MIT MIT License
- * @copyright 2017 Ronan GIRON
+ * @copyright 2021 Ronan GIRON
  * @author    Ronan GIRON <https://github.com/ElGigi>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code, to the root.
  */
-
-putenv('BERLIOZ_FOO=BAR');
 
 $file = __DIR__ . '/../vendor/autoload.php';
 if (!file_exists($file)) {
